+++ app/dashboard/properties/[propertyId]/delete/_components/DeletePropertyForm.tsx
@@ -1,26 +1,59 @@
 'use client';
 
-import { ObjectDeletionForm } from '@/components/New/Forms/ObjectDeletionForm';
-import { MainHeading } from '@/components/New/Typography/Headings';
 import { FormControl, Input } from '@/components/UI/FormUtils';
 import { AppartmentPayloadType, HousePayloadType } from 'kotilogi-app/dataAccess/types';
 import { deletePropertyAction } from './actions';
 import { BoxFieldset } from '@/components/UI/BoxFieldset';
+import toast from 'react-hot-toast';
+import { PropertyError, UserError } from 'kotilogi-app/utils/error';
+import { useRouter } from 'next/navigation';
+import { useState } from 'react';
+import { useFormOnChangeObject } from '@/hooks/useFormOnChangeObject';
+import { FormButtons } from '@/components/New/Forms/FormBase';
 
 type DeletePropertyFormProps = {
   property: HousePayloadType | AppartmentPayloadType;
 };
 
 export function DeletePropertyForm({ property }: DeletePropertyFormProps) {
+  const router = useRouter();
+  const [status, setStatus] = useState<'idle' | 'loading' | 'done' | 'error'>('idle');
+  const { data: credentials, updateData: updateCredentials } = useFormOnChangeObject({
+    password: '',
+  });
+
+  const deleteProperty = async (e: any) => {
+    e.preventDefault();
+    let currentStatus: typeof status = 'loading';
+    setStatus(currentStatus);
+    try {
+      const result = await deletePropertyAction(property.id, credentials.password);
+      if (result === UserError.INVALID_PASSWORD) {
+        toast.error('Salasana on virheellinen!');
+        currentStatus = 'error';
+      } else if (result === UserError.NOT_OWNER) {
+        toast.error('Taloa ei voi poistaa, sillä et ole sen omistaja!');
+        currentStatus = 'error';
+      } else if (result === PropertyError.OPEN_TRANSFERS) {
+        toast.error('Taloa ei voi poistaa, sillä se odottaa siirtoa toiselle käyttäjälle.');
+        currentStatus = 'error';
+      } else {
+        toast.success('Talo poistettu!');
+        router.replace('/dashboard/properties');
+        currentStatus = 'done';
+      }
+    } catch (err) {
+      toast.error('Talon poisto epäonnistui tuntemattomasta syystä!');
+      currentStatus = 'error';
+    } finally {
+      setStatus(currentStatus);
+    }
+  };
+
   return (
     <BoxFieldset legend='Poista talo'>
       <div className='flex flex-col gap-4 w-full'>
-        <ObjectDeletionForm
-          returnUrl='/dashboard'
-          objectId={property.id}
-          deleteMethod={async credentials =>
-            deletePropertyAction(property.id, credentials.password)
-          }>
+        <form onSubmit={deleteProperty}>
           <p className='text-lg mb-4'>
             Olet poistamassa taloa{' '}
             <strong className='font-semibold'>
@@ -43,10 +76,18 @@ export function DeletePropertyForm({ property }: DeletePropertyFormProps) {
                 name='password'
                 autoComplete='new-password'
                 placeholder='Kirjoita salasanasi...'
+                onChange={updateCredentials}
+                value={credentials.password}
               />
             }
           />
-        </ObjectDeletionForm>
+          <FormButtons
+            loading={status === 'loading'}
+            done={status === 'done'}
+            submitDisabled={credentials.password.length === 0}
+            backAction={() => router.back()}
+          />
+        </form>
       </div>
     </BoxFieldset>
   );