+++ app/dashboard/properties/[propertyId]/events/add/_components/FormContent/HeatingRenovationContent/HeatingSystemSelector.tsx
@@ -17,6 +17,7 @@ export const HeatingSystemSelector = ({
 }) => {
   const { payload, updatePayload, resetPayload, editing } = useEventFormContext();
   useEffect(() => {
+    //console.log('Selected heating type: ', payload.new_heating_type);
     /*Reset the rest of the payload if the new system is changed.
     console.log('Resetting payload...');
     resetPayload({
@@ -42,7 +43,10 @@ export const HeatingSystemSelector = ({
                 key={`heating_type_id-${i}`}
                 checked={payload.new_heating_type == t}
                 name={'new_heating_type'}
-                onChange={updatePayload}
+                onChange={e => {
+                  console.log(e.target.value);
+                  updatePayload(e);
+                }}
               />
             );
           })}