+++ utils/error.ts
@@ -1,0 +1,8 @@
+export enum UserError {
+  INVALID_PASSWORD = 'invalid_password',
+  NOT_OWNER = 'not_owner',
+}
+
+export enum PropertyError {
+  OPEN_TRANSFERS = 'open_transfers',
+}